ignore SIGTERM and SIGINT on teardown
to make the teardown more robust and not fail half way ignore common
kill signals, SIGTERM seems be send to us by systemd under some
circumstances.
It is far from perfect and not fixing the root cause, somthing seem to
be destroying the cgroup before we are done with cleanup in podman.

Fixes #1223

Signed-off-by: Paul Holzinger <pholzing@redhat.com>