remove some unused deps
So cargo doesn't really care if we list stuff there that we don't use.

Neither serde-value nor netlink-packet-utils were used directly. Also
putting the json deps behind a feature makes no sense as we always need
that to build and we ar enot a lib. Removing the url feature allos us to
drop the url dep as well which we also do not use.

And then while I touch the serde json deps update them to the latest.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>