aardvark: handle names more safely
Since we write space, command and newline separated file these
characters have a special meaning when parsing the file. Any such
character inside the strings can break the parser. We need checks to
disallow these characters.

While it would have been easy to add a check at the time we write the
file there we could not really make a good decision on what to do with
the value. As such I create a newtype which wraps the string but it can
only be constructed via the TryFrom trait. That is enough to gurantee
that any caller must have called the validation code.

Note that we only print a warning and ignore such invalid value, a hard
error might cause compatibility issues as per the issue some automated
tooling generates such names.

Fixes: #1019
Fixes: containers/podman#25927

Signed-off-by: Paul Holzinger <pholzing@redhat.com>