drop sysctl dependency
We really don't need it, we can open/read/write the files directly which
is not more complex. The one thing we don't do is accept the path in any
format but I see this an benifit we should be consitent in how we pass
in the path name.

This also removed one extra open call on the file so this is better.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>