write bridge sysctl to config file
When writing directly to /proc/sys we have the risk that the system
later overwrites the values again, for systemd distros systemd-sysctl
seems to manage that and reapplies sysctls based of the sysctl.d config
files.

Now as RHEL ships a config file that set rp_filter = 1 by default this
means that our rp_filter = 2 could be overwritten again by it.
So in short the original fix to set rp_filter = 2 never worked actually
as the value got overwritten again.

So to actually fix this create a sysctl.d config file for the bridge
with all its required sysctl settings. However we connot only do that,
as this happens async systemd might have no set the value by the time we
start the container process. So to avoid any reaces we do both.
Also of course we should only do that when systemd is in use and when
run as root, as rootless we operate in our own private netns so systemd
doesn't see the interfaces there and we do not have the privileges to
write to /run either.

Fixes: 9a73a52 ("fix port forward with strict RPF and multi networks")
Fixes: https://issues.redhat.com/browse/RHEL-89477

Signed-off-by: Paul Holzinger <pholzing@redhat.com>