sysctl: restructure bridge sysctl settings
First always set ipv4/6 global forwarding and don't use a sync once,
when multiple networks are in use we do set them again but we now want
them to write them into the per bridge sysctl.d config file.

That way we can make these sysctls track for all running bridges and
once all contianer are stopped we do not end up with leaked sysctl.d
file which seems nice. And we also only have to handle one file at a
time.

Then moce the internal forwarding disable right next to the global
enable and only enable it globally when not internal.
Lastly always set the route_localnet even when no ports are given.
The reason is we now set the sysctl only once when we create the bridge.
This offers the advantage that when starting many containers on the same
network we only have to configure the bridge specifc sysctl values once.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>