drop DhcpProxy error type from NetavarkError
The tonic status is quite large and thus a new clippy lint now seem to
fail[1].

Only the test binary used the variant anyway so just drop it and also
print only the error text as the metadata is not useful in the error
message. Then update the tests to check the message instead of exit
code.

[1] https://rust-lang.github.io/rust-clippy/master/index.html#result_large_err

Signed-off-by: Paul Holzinger <pholzing@redhat.com>