add .cargo/config.toml to runs tests via unshare -rn
Out test don't actually need root, we only care about the ability to
create a private netns in src/test/netlink.rs. This can be done via a
user namespace as well. So configure cargo to run the tests via
unshare -rn which creates a private userns and netns. That way devs can
run cargo test rootless on the system.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>