test: add new test program for connection checks
Use this to replace ncat. ncat is flaky and harder to use and with
the latest version it seems they broke us in a way that makes the flags
we use incompatible[1].

As a nice benefit this makes the test code itself simpler and faster as
we get rid of the is port bound polling/sleeps.

[1] https://github.com/containers/netavark/pull/1282

Fixes: #1076

Signed-off-by: Paul Holzinger <pholzing@redhat.com>