bridge: bind ip for aardvark-dns in unmanaged mode if gateway ip is not on the host
Find the Universe scope IPv4 addresses of the bridge with the modified dump_addresses() function.
This function works for all interfaces, not just bridges.

If the dns is enabled and the bridge mode is unmanaged, then the bind IP of aardvark-dns is changed to the IP addresses of the bridge instead of the gateway. If there are no IP address on the bridge then we just fail with a clear error that the user must disable dns (--disable-dns) when creating the network.

Signed-off-by: Shivang K Raghuvanshi <shivangraghuvanshi2005@gmail.com>