bridge: bind ip for aardvark-dns in unmanaged mode if gateway ip is not on the host
Find all the IPv4 and IPv6 addresses except link local IPv6 addressses of the bridge with the modified dump_addresses() function.

If the dns is enabled and the bridge mode is unmanaged, then the bind IP of aardvark-dns is changed to the IP addresses of the bridge instead of the gateway. If there is no usable IP address on the bridge then we just fail with a clear error that the user must disable dns (--disable-dns) when creating the network.

Fixes: #1177
Signed-off-by: Shivang K Raghuvanshi <shivangraghuvanshi2005@gmail.com>