Packit: fetch copr rpm version from Cargo.toml
Current installation of nv from podman-next is failing because it's
expecting aardvark-dns >= 102:1.16 while current av on podman-next is
only 102:1.15 . From the latest build log, that seems to be because
Packit is determining 1.15.0 as the version for av and 1.16.0 as the
version for nv.

This commit includes a script to fetch the version directly from Cargo.toml. A
similar script will be added to av, thus ensuring both nv and av
major.minor versions are in sync for copr rpms as well.

Related: https://github.com/containers/aardvark-dns/pull/625/

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>