make network setup order deterministic
Change the file format of the networks json to use an array for the
PerNetworkOptions. With that we can get a defined order in which we
setup the networks and do that deterministically.

In order to provide backwards compatibility we use a custom json
deserializer function that accepts both the old and new format.
That allows us to still use the old podman against the new netavark
with this patch.

I also added PartialEq trait to the types so the unit test can compare
two loaded Networks.

I only converted a few of our test configs to the new format to keep the
diff smaller. We can convert most of them another time.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>